requirements.txt
six
pytz
coreapi>=1.32
django>=1.10
djangorestframework>=3.4
dynamicforms<0.70.0
typing; python_version < '3.5'
